Extract liquidation to sperate crate.
