+++ .github/workflows/coverage.yml
@@ -25,10 +25,6 @@ jobs:
       - name: Install stable Rust
         run: cargo make install-stable
 
-      # selecting a toolchain should happen before the plugin, as the cache uses the current rustc version as its cache key
-      # - name: Cache dependencies
-      #   uses: Swatinem/rust-cache@v2
-
       # Artifacts used by tests
       - name: Compile workspace
         run: cargo make build
