+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "contracts/rewards-collector",
   "packages/chains/*",
   "packages/health",
+  "packages/liquidation",
   "packages/testing",
   "packages/types",
   "packages/utils",
@@ -66,6 +67,7 @@ proptest          = "1.1.0"
 
 # packages
 mars-health         = { version = "1.0.0", path = "./packages/health" }
+mars-liquidation    = { version = "1.0.0", path = "./packages/liquidation" }
 mars-osmosis        = { version = "1.0.0", path = "./packages/chains/osmosis" }
 mars-params         = { version = "1.0.7", path = "./contracts/params" }
 mars-red-bank-types = { version = "1.0.0", path = "./packages/types" }
