+++ contracts/red-bank/Cargo.toml
@@ -27,6 +27,7 @@ cw2                 = { workspace = true }
 cw-storage-plus     = { workspace = true }
 cw-utils            = { workspace = true }
 mars-health         = { workspace = true }
+mars-liquidation    = { workspace = true }
 mars-owner          = { workspace = true }
 mars-params         = { workspace = true }
 mars-red-bank-types = { workspace = true }
