+++ packages/liquidation/README.md
@@ -1,0 +1,7 @@
+# Mars Health
+
+Functions used for evaluating the liquidation of user positions at Mars Protocol.
+
+## License
+
+Contents of this crate are open source under [GNU General Public License v3](../../LICENSE) or later.
