+++ packages/liquidation/src/error.rs
@@ -1,0 +1,23 @@
+use cosmwasm_std::{
+    CheckedFromRatioError, CheckedMultiplyFractionError, CheckedMultiplyRatioError, OverflowError,
+    StdError,
+};
+use thiserror::Error;
+
+#[derive(Error, Debug, PartialEq)]
+pub enum LiquidationError {
+    #[error("{0}")]
+    Std(#[from] StdError),
+
+    #[error("{0}")]
+    Overflow(#[from] OverflowError),
+
+    #[error("{0}")]
+    CheckedMultiplyRatio(#[from] CheckedMultiplyRatioError),
+
+    #[error("{0}")]
+    CheckedMultiplyFraction(#[from] CheckedMultiplyFractionError),
+
+    #[error("{0}")]
+    CheckedFromRatio(#[from] CheckedFromRatioError),
+}
